+++ client-common/src/main/java/org/realityforge/replicant/client/EntitySystem.java
@@ -1,0 +1,18 @@
+package org.realityforge.replicant.client;
+
+import javax.annotation.Nonnull;
+
+/**
+ * The service that exposes the EntityRepository, the EntityChangeBroker and the EntitySubscriptionManager.
+ */
+public interface EntitySystem
+{
+  @Nonnull
+  EntityRepository getRepository();
+
+  @Nonnull
+  EntityChangeBroker getChangeBroker();
+
+  @Nonnull
+  EntitySubscriptionManager getSubscriptionManager();
+}