+++ client-gwt/src/main/java/org/realityforge/replicant/client/gwt/ReplicantGinModule.java
@@ -1,12 +1,13 @@
 package org.realityforge.replicant.client.gwt;
 
 import com.google.gwt.inject.client.AbstractGinModule;
+import com.google.inject.Provides;
+import javax.inject.Singleton;
 import org.realityforge.replicant.client.EntityChangeBroker;
-import org.realityforge.replicant.client.EntityChangeBrokerImpl;
 import org.realityforge.replicant.client.EntityRepository;
-import org.realityforge.replicant.client.EntityRepositoryImpl;
 import org.realityforge.replicant.client.EntitySubscriptionManager;
-import org.realityforge.replicant.client.EntitySubscriptionManagerImpl;
+import org.realityforge.replicant.client.EntitySystem;
+import org.realityforge.replicant.client.EntitySystemImpl;
 
 /**
  * A simple GIN module that defines the repository and change broker services.
@@ -17,8 +18,27 @@ public class ReplicantGinModule
   @Override
   protected void configure()
   {
-    bind( EntityRepository.class ).to( EntityRepositoryImpl.class ).asEagerSingleton();
-    bind( EntityChangeBroker.class ).to( EntityChangeBrokerImpl.class ).asEagerSingleton();
-    bind( EntitySubscriptionManager.class ).to( EntitySubscriptionManagerImpl.class ).asEagerSingleton();
+    bind( EntitySystem.class ).to( EntitySystemImpl.class ).asEagerSingleton();
+  }
+
+  @Provides
+  @Singleton
+  public final EntityRepository getEntityRepository( final EntitySystem system )
+  {
+    return system.getRepository();
+  }
+
+  @Provides
+  @Singleton
+  public final EntityChangeBroker getEntityChangeBroker( final EntitySystem system )
+  {
+    return system.getChangeBroker();
+  }
+
+  @Provides
+  @Singleton
+  public final EntitySubscriptionManager getEntitySubscriptionManager( final EntitySystem system )
+  {
+    return system.getSubscriptionManager();
   }
 }